requirements.txt
Django==4.1.5
djangorestframework==3.14.0
django-cors-headers==3.13.0
psycopg2==2.9.5
gunicorn==20.1.0
whitenoise==6.3.0
dj-database-url==1.2.0